requirements.txt
langchain
python-dotenv
google-generativeai
langchain-google-genai
PyPDF2
google-generativeai
reportlab
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
streamlit
pymupdf
